# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Delivery service",
        "title": "DeliVeryWell",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/",
    "paths": {
        "/fines": {
            "get": {
                "description": "Get list of all fines or search fines by keyword",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Get all fines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search fines",
                        "name": "searchFines",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FinesListWithRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new fine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Create a new fine",
                "parameters": [
                    {
                        "description": "Fine data",
                        "name": "fine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Fines"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ds.Fines"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines-resolutions/{id}": {
            "delete": {
                "description": "Delete fine-resolution link by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines-Resolutions"
                ],
                "summary": "Delete fine-resolution link by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine-Resolution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines-resolutions/{id}/count": {
            "put": {
                "description": "Update the count of a fine-resolution link",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines-Resolutions"
                ],
                "summary": "Update fine-resolution count",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine-Resolution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fine-Resolution data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Fine_Resolutions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Fine_Resolutions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines/{id}": {
            "get": {
                "description": "Get a fine by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Get fine by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Fines"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update fine details by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Update a fine by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fine data",
                        "name": "fine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Fines"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Fines"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete fine by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Delete a fine by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines/{id}/resolution": {
            "post": {
                "description": "Add fine to a user's resolution",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fines"
                ],
                "summary": "Add a fine to a resolution",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/images/{id}": {
            "post": {
                "description": "Upload an image for a specific ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Upload an image by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ImageURL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Fine_Resolutions": {
            "type": "object",
            "properties": {
                "fin_Res_ID": {
                    "type": "integer"
                },
                "fine_ID": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "resolution_ID": {
                    "type": "integer"
                }
            }
        },
        "ds.Fines": {
            "type": "object",
            "properties": {
                "dopInf": {
                    "type": "string"
                },
                "fineID": {
                    "type": "integer"
                },
                "fullInf": {
                    "type": "string"
                },
                "imge": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.FinesListWithRes": {
            "type": "object",
            "properties": {
                "fines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Fines"
                    }
                },
                "resCount": {
                    "type": "integer"
                },
                "resId": {
                    "type": "integer"
                }
            }
        }
    }
}